Move embedding types into metabase-types
