+++ enterprise/frontend/src/metabase-enterprise/resource_downloads/resource_downloads_plugin.ts
@@ -1,8 +1,8 @@
 import { P, match } from "ts-pattern";
 
-import type { EmbedResourceDownloadOptions } from "metabase/embedding/types";
 import { PLUGIN_RESOURCE_DOWNLOADS } from "metabase/plugins";
 import { hasPremiumFeature } from "metabase-enterprise/settings";
+import type { EmbedResourceDownloadOptions } from "metabase-types/api";
 
 /**
  * Initialize resource downloads plugin features that depend on hasPremiumFeature.
