+++ frontend/src/metabase-types/api/index.ts
@@ -21,6 +21,7 @@ export * from "./dataset";
 export * from "./dependencies";
 export * from "./document";
 export * from "./email";
+export * from "./embed";
 export * from "./embedding-theme";
 export * from "./entity-id";
 export * from "./field";
