+++ frontend/src/metabase/dashboard/types/display-options.ts
@@ -1,4 +1,4 @@
-import type { EmbedResourceDownloadOptions } from "metabase/embedding/types";
+import type { EmbedResourceDownloadOptions } from "metabase-types/api";
 
 export type DashboardFullscreenControls = {
   isFullscreen: boolean;
