+++ frontend/src/metabase/dashboard/types/embed-display-options.ts
@@ -1,8 +1,6 @@
-import type {
-  DisplayTheme,
-  EmbedResourceDownloadOptions,
-} from "metabase/embedding/types";
+import type { DisplayTheme } from "metabase/embedding/types";
 import type { ClickActionModeGetter } from "metabase/visualizations/types";
+import type { EmbedResourceDownloadOptions } from "metabase-types/api";
 
 type EmbedBackground = boolean;
