+++ frontend/src/metabase/embedding/components/EmbedModal/StaticEmbedSetupPane/ParametersSettings.tsx
@@ -8,14 +8,16 @@ import CS from "metabase/css/core/index.css";
 import type {
   EmbedResourceParameter,
   EmbedResourceType,
-  EmbeddingParameterVisibility,
-  EmbeddingParameters,
   EmbeddingParametersValues,
 } from "metabase/embedding/types";
 import { ParameterWidget as StaticParameterWidget } from "metabase/parameters/components/ParameterWidget";
 import { getParameterIconName } from "metabase/parameters/utils/ui";
 import { Box, Divider, Icon, Stack, Text } from "metabase/ui";
 import { getValuePopulatedParameters } from "metabase-lib/v1/parameters/utils/parameter-values";
+import type {
+  EmbeddingParameterVisibility,
+  EmbeddingParameters,
+} from "metabase-types/api";
 
 import { StaticEmbedSetupPaneSettingsContentSection } from "./StaticEmbedSetupPaneSettingsContentSection";
 import type { EmbedResourceParameterWithValue } from "./types";
