+++ frontend/src/metabase/embedding/constants.ts
@@ -1,4 +1,4 @@
-import type { EmbeddingType } from "metabase/embedding/types";
+import type { EmbeddingType } from "metabase-types/api";
 
 export const STATIC_LEGACY_EMBEDDING_TYPE: EmbeddingType = "static-legacy";
 export const GUEST_EMBED_EMBEDDING_TYPE: EmbeddingType = "guest-embed";
