+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/analytics.ts
@@ -8,9 +8,8 @@ import type {
 import { getAuthSubTypeForSettings } from "metabase/embedding/embedding-iframe-sdk-setup/utils/get-auth-sub-type-for-settings";
 import { getAuthTypeForSettings } from "metabase/embedding/embedding-iframe-sdk-setup/utils/get-auth-type-for-settings";
 import { countEmbeddingParameterOptions } from "metabase/embedding/lib/count-embedding-parameter-options";
-import type { EmbeddingParameters } from "metabase/embedding/types";
 import type { SdkIframeEmbedSetupModalInitialState } from "metabase/plugins";
-import type { Card, Dashboard } from "metabase-types/api";
+import type { Card, Dashboard, EmbeddingParameters } from "metabase-types/api";
 
 import type {
   SdkIframeEmbedSetupExperience,
