+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/context.ts
@@ -1,11 +1,13 @@
 import { createContext, useContext } from "react";
 
-import type {
-  EmbeddingParameters,
-  EmbeddingParametersValues,
-} from "metabase/embedding/types";
+import type { EmbeddingParametersValues } from "metabase/embedding/types";
 import type { SdkIframeEmbedSetupModalInitialState } from "metabase/plugins";
-import type { Card, Dashboard, Parameter } from "metabase-types/api";
+import type {
+  Card,
+  Dashboard,
+  EmbeddingParameters,
+  Parameter,
+} from "metabase-types/api";
 
 import type {
   SdkIframeEmbedSetupExperience,
