+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/hooks/use-embedding-parameters-conversion.ts
@@ -1,8 +1,7 @@
 import { useCallback } from "react";
 
 import type { SdkIframeEmbedSetupSettings } from "metabase/embedding/embedding-iframe-sdk-setup/types";
-import type { EmbeddingParameters } from "metabase/embedding/types";
-import type { Parameter } from "metabase-types/api";
+import type { EmbeddingParameters, Parameter } from "metabase-types/api";
 
 export const useEmbeddingParametersConversion = () => {
   const convertToEmbedSettings = useCallback(
