+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/hooks/use-parameters-values.ts
@@ -3,8 +3,7 @@ import { useMemo } from "react";
 import { getPreviewParamsBySlug } from "metabase/embedding/components/EmbedModal/StaticEmbedSetupPane/lib/get-preview-params-by-slug";
 import type { SdkIframeEmbedSetupSettings } from "metabase/embedding/embedding-iframe-sdk-setup/types";
 import { convertParameterValuesBySlugToById } from "metabase/embedding/embedding-iframe-sdk-setup/utils/convert-parameter-values-by-slug-to-by-id";
-import type { EmbeddingParameters } from "metabase/embedding/types";
-import type { Parameter } from "metabase-types/api";
+import type { EmbeddingParameters, Parameter } from "metabase-types/api";
 
 interface UseParametersValuesProps {
   settings: SdkIframeEmbedSetupSettings;
