+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/hooks/use-toggle-resource-embedding.ts
@@ -7,7 +7,7 @@ import {
 import { useSetting } from "metabase/common/hooks";
 import { getStaticEmbedSetupPublishHandlers } from "metabase/embedding/components/EmbedModal/StaticEmbedSetupPane/lib/get-static-embed-setup-publish-handlers";
 import { GUEST_EMBED_EMBEDDING_TYPE } from "metabase/embedding/constants";
-import type { EmbeddingParameters } from "metabase/embedding/types";
+import type { EmbeddingParameters } from "metabase-types/api";
 
 import { useSdkIframeEmbedSetupContext } from "../context";
 import { getResourceTypeFromExperience } from "../utils/get-resource-type-from-experience";
