+++ frontend/src/metabase/query_builder/components/template_tags/TagEditorSidebar.tsx
@@ -3,13 +3,13 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import { SidebarContent } from "metabase/common/components/SidebarContent";
-import type { EmbeddingParameterVisibility } from "metabase/embedding/types";
 import { Box, Tabs } from "metabase/ui";
 import type Question from "metabase-lib/v1/Question";
 import type Database from "metabase-lib/v1/metadata/Database";
 import type NativeQuery from "metabase-lib/v1/queries/NativeQuery";
 import type {
   DatabaseId,
+  EmbeddingParameterVisibility,
   NativeDatasetQuery,
   Parameter,
   ParameterId,
